+++ src/Brp.Shared.Infrastructure/ProblemDetails/ProblemDetailsFactory.cs
@@ -33,9 +33,12 @@ public static class ProblemDetailsFactory
             StatusCodes.Status400BadRequest => "Een of meerdere parameters zijn niet correct.",
             StatusCodes.Status401Unauthorized => "Niet correct geauthenticeerd.",
             StatusCodes.Status403Forbidden => "U bent niet geautoriseerd voor het gebruik van deze API.",
+            StatusCodes.Status404NotFound => "Opgevraagde resource bestaat niet.",
             StatusCodes.Status405MethodNotAllowed => "Gebruikte bevragingsmethode is niet toegestaan.",
             StatusCodes.Status406NotAcceptable => "Gevraagde content type wordt niet ondersteund.",
             StatusCodes.Status415UnsupportedMediaType => "Media Type wordt niet ondersteund.",
+            StatusCodes.Status500InternalServerError => "Internal Server error.",
+            StatusCodes.Status502BadGateway => "Bronservice is tijdelijk niet beschikbaar.",
             _ => null
         };
 
@@ -45,9 +48,12 @@ public static class ProblemDetailsFactory
             StatusCodes.Status400BadRequest => StatusCodeIdentifiers.BadRequestIdentifier,
             StatusCodes.Status401Unauthorized => StatusCodeIdentifiers.UnauthorizedIdentifier,
             StatusCodes.Status403Forbidden => StatusCodeIdentifiers.ForbiddenIdentifier,
+            StatusCodes.Status404NotFound => StatusCodeIdentifiers.NotFoundIdentifier,
             StatusCodes.Status405MethodNotAllowed => StatusCodeIdentifiers.MethodNotAllowedIdentifier,
             StatusCodes.Status406NotAcceptable => StatusCodeIdentifiers.NotAcceptableIdentifier,
             StatusCodes.Status415UnsupportedMediaType => StatusCodeIdentifiers.UnsupportedMediaTypeIdentifier,
+            StatusCodes.Status500InternalServerError => StatusCodeIdentifiers.InternalServerErrorIdentifier,
+            StatusCodes.Status502BadGateway => StatusCodeIdentifiers.ServiceUnavailableIdentifier,
             _ => null
         };