+++ src/Brp.Shared.Infrastructure/ProblemDetails/StatusCodeIdentifiers.cs
@@ -5,8 +5,10 @@ public static class StatusCodeIdentifiers
     public const string BadRequestIdentifier = "https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.1";
     public const string UnauthorizedIdentifier = "https://datatracker.ietf.org/doc/html/rfc7235#section-3.1";
     public const string ForbiddenIdentifier = "https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.3";
+    public const string NotFoundIdentifier = "https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.4";
     public const string MethodNotAllowedIdentifier = "https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.5";
     public const string NotAcceptableIdentifier = "https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.6";
     public const string UnsupportedMediaTypeIdentifier = "https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.13";
     public const string InternalServerErrorIdentifier = "https://datatracker.ietf.org/doc/html/rfc7231#section-6.6.1";
+    public const string ServiceUnavailableIdentifier = "https://datatracker.ietf.org/doc/html/rfc7231#section-6.6.4";
 }